move service classes to services module
