+++ conveyor/src/main/kotlin/com/hodamohammadi/conveyor/activities/AuthenticationActivity.kt
@@ -7,8 +7,8 @@ import android.widget.Toast
 import com.hodamohammadi.conveyor.R
 import com.firebase.ui.auth.AuthUI
 import com.firebase.ui.auth.IdpResponse
-import com.hodamohammadi.commons.navigation.loadIntentOrNull
-import com.hodamohammadi.conveyor.services.RoutePath
+import com.hodamohammadi.commons.navigation.loaders.loadIntentOrNull
+import com.hodamohammadi.commons.navigation.RoutePath
 import java.util.*
 
 /**
