+++ navigation/src/main/kotlin/com/hodamohammadi/commons/navigation/loaders/IntentLoader.kt
@@ -1,4 +1,4 @@
-package com.hodamohammadi.commons.navigation
+package com.hodamohammadi.commons.navigation.loaders
 
 import android.content.Intent
